+++ pkg/sessions/redis/session_store_test.go
@@ -1,0 +1,47 @@
+package redis
+
+import (
+	"testing"
+	"time"
+
+	"github.com/alicebob/miniredis/v2"
+	"github.com/oauth2-proxy/oauth2-proxy/pkg/apis/options"
+	sessionsapi "github.com/oauth2-proxy/oauth2-proxy/pkg/apis/sessions"
+	"github.com/oauth2-proxy/oauth2-proxy/pkg/logger"
+	"github.com/oauth2-proxy/oauth2-proxy/pkg/sessions/tests"
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestSessionStore(t *testing.T) {
+	logger.SetOutput(GinkgoWriter)
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Redis SessionStore")
+}
+
+var _ = Describe("Redis SessionStore Tests", func() {
+	var mr *miniredis.Miniredis
+
+	BeforeEach(func() {
+		var err error
+		mr, err = miniredis.Run()
+		Expect(err).ToNot(HaveOccurred())
+	})
+
+	AfterEach(func() {
+		mr.Close()
+	})
+
+	tests.RunSessionStoreTests(
+		func(opts *options.SessionOptions, cookieOpts *options.CookieOptions) (sessionsapi.SessionStore, error) {
+			// Set the connection URL
+			opts.Type = options.RedisSessionStoreType
+			opts.Redis.ConnectionURL = "redis://" + mr.Addr()
+			return NewRedisSessionStore(opts, cookieOpts)
+		},
+		func(d time.Duration) error {
+			mr.FastForward(d)
+			return nil
+		},
+	)
+})